+++ mountpoint-s3-fs/src/prefetch/part_queue.rs
@@ -214,7 +214,7 @@ mod tests {
                     let body: Box<[u8]> = (0u8..=255).cycle().skip(offset as u8 as usize).take(n).collect();
                     let bytes: Bytes = body.into();
                     let checksummed_bytes = ChecksummedBytes::new(bytes);
-                    let part = Part::new(part_id.clone(), offset, checksummed_bytes);
+                    let part = Part::new(part_id.clone(), offset, checksummed_bytes, false);
                     part_queue_producer.push(Ok(part));
                     current_length += n;
                 }